@@ -7,37 +7,32 @@
     "deploy": "npm run build && npm run server",
     "build": "npm run clean && node ./scripts/prodBundler.js",
     "server": "node ./scripts/server.js",
-    "test": "jest — coverage — no-cache ${1}",
-    "test-watch": "jest ${1} — watch",
+    "test": "jest --coverage --no-cache",
+    "test-watch": "jest --watch",
     "clean": "node ./scripts/clean.js",
     "install-utils": "typings install && yarn install -g && yarn install",
     "start": "npm run clean && node ./scripts/devBundler.js && concurrently \"node ./scripts/hotReload.js\" \"nodemon ./scripts/server.js --watch public -e js\" --kill-others",
+    "start-new": "npm run clean && node ./scripts/devBundler.js && concurrently \"node ./scripts/fuseDev.js\" \"nodemon ./scripts/server.js --watch public -e js\" --kill-others",
     "lint": "tslint './**/*.ts{,x}'"
   },
   "author": "Will McVay",
   "license": "MIT",
   "jest": {
+    "transform": {
+      ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+    },
+    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
     "moduleFileExtensions": [
       "ts",
+      "tsx",
       "js"
-    ],
-    "transform": {
-      "\\.(ts)$": "<rootDir>/node_modules/ts-jest/preprocessor.js"
-    },
-    "testRegex": "lib/.*\\.spec.(ts|js)$",
-    "testResultsProcessor": "<rootDir>/node_modules/ts-jest/coverageprocessor.js",
-    "globals": {
-      "__TS_CONFIG__": {
-        "target": "es6",
-        "module": "commonjs",
-        "moduleResolution": "node"
-      }
-    }
+    ]
   },
   "dependencies": {
     "axios": "^0.15.2",
     "compression": "^1.6.2",
     "express": "^4.13.3",
+    "fs": "^0.0.1-security",
     "history": "^4.4.0",
     "mongodb": "^2.0.49",
     "mongoose": "^4.2.8",
@@ -51,6 +46,7 @@
     "socket.io": "^1.3.7"
   },
   "devDependencies": {
+    "@types/jest": "^19.2.2",
     "awesome-typescript-loader": "^3.0.0-beta.18",
     "concurrently": "^3.1.0",
     "css-loader": "^0.26.1",
@@ -58,7 +54,8 @@
     "enzyme": "~2.6.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
-    "fs": "0.0.1-security",
+    "fuse-box": "^1.3.130",
+    "gulp": "^3.9.1",
     "jasmine-core": "^2.5.2",
     "jest": "^18.1.0",
     "node-sass": "^3.13.0",
@@ -69,9 +66,9 @@
     "redux-logger": "^2.7.4",
     "sass-loader": "^4.1.1",
     "style-loader": "^0.13.1",
-    "ts-jest": "^18.0.1",
+    "ts-jest": "^18.0.3",
     "tslint": "^4.0.1",
-    "typescript": "^2.1.0",
+    "typescript": "^2.2.2",
     "typings": "^2.1.0",
     "webpack": "~1.13.2",
     "webpack-dev-server": "~1.16.2"
