@@ -2,6 +2,7 @@ const webpack          = require( 'webpack' );
 const WebpackDevServer = require( 'webpack-dev-server' );
 const configClient     = require( './webpack.dev.client.js' );
 const configServer     = require( './webpack.dev.server.js' );
+const fsbx = require('fuse-box');
 
 const options = {
   chunk       : false,
@@ -12,15 +13,23 @@ const options = {
   quiet       : true
 };
 
-new WebpackDevServer( webpack( configClient ), {
-  hot: true,
-  historyApiFallback: true,
-  stats: options
-}).listen( 8000, 'localhost', ( err ) => {
-  if(err) console.log( `Webpack client bundle error: ${ err }` );
+// new WebpackDevServer( webpack( configClient ), {
+//   hot: true,
+//   historyApiFallback: true,
+//   stats: options
+// }).listen( 8000, 'localhost', ( err ) => {
+//   if(err) console.log( `Webpack client bundle error: ${ err }` );
 
-  console.log( 'Webpack client server launched with at localhost:8000 [HMR] enabled)' );
-});
+//   console.log( 'Webpack client server launched with at localhost:8000 [HMR] enabled)' );
+// });
+
+
+
+new fsbx.FuseBox({
+  homeDir: "src/",
+  outFile: "public/client.js",
+  sourceMaps: true
+}).devServer(">client/index.tsx");
 
 webpack( configServer ).watch( {}, ( err, stats ) => {
   if( err ) return console.error( `Webpack server bundle error: ${ err.message }` );
